requirements.txt
bleach==1.5.0
colorama==0.3.9
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
html5lib==0.9999999
imgaug==0.2.4
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
notebook==5.0.0
numpy==1.13.1+mkl
olefile==0.44
opencv-python==3.3.0
pandas==0.20.3
pandoc==1.0.2
pandocfilters==1.4.2
pickleshare==0.7.4
Pillow==4.1.1
ply==3.10
prettyprint==0.1.5
prompt-toolkit==1.0.15
protobuf==3.4.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
pyzmq==16.0.2
qtconsole==4.3.1
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.2
win-unicode-console==0.5
